requirements.txt
streamlit
streamlit-option-menu
python-dotenv
fpdf
google-generativeai
